refactor: improve dataset download and extraction logic

+++ Dataset_import.py
@@ -1,14 +1,14 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-# # Step 1.  
+# # Step 1.
 # #### Dataset download, import and decompression
-# The datasets are downloaded using wget and decompressed using tar.  
-# The file datasets.json contains the url of all the task datasets and names.  
-# The following loop iters over the json file to download all the datasets.  
-# 
-# 
-# wget {url} -O NAME  
+# The datasets are downloaded using wget and decompressed using tar.
+# The file datasets.json contains the url of all the task datasets and names.
+# The following loop iters over the json file to download all the datasets.
+#
+#
+# wget {url} -O NAME
 # tar -zxf NAME
 
 # In[ ]:
@@ -31,18 +31,22 @@ os.makedirs(log_dir, exist_ok=True)
 os.makedirs(image_dir, exist_ok=True)
 
 # Redirect stdout and stderr to a log file
-log_file = open(f"{log_dir}/{script_name}_{timestamp}.log", 'w')
+log_file = open(f"{log_dir}/{script_name}_{timestamp}.log", "w")
 sys.stdout = log_file
 sys.stderr = log_file
+
 
 # Ensure log file is closed on exit
 def close_log():
     log_file.close()
+
+
 import atexit
+
 atexit.register(close_log)
 
 # Read datasets information from JSON
-with open('datasets.json', 'r') as file:
+with open("datasets.json", "r") as file:
     datasets = json.load(file)
 
 # Perform downloads
@@ -53,20 +57,21 @@ for task in datasets:
 
     Path("tasks_compressed").mkdir(exist_ok=True)
 
-    if not(os.path.exists(f"tasks/{task_name}")):
-        print(f"Downloading: {task_name}")
-        
-        # Download
-        os.system(f"wget {task_url} -O tasks_compressed/{task_name}")
-    
-        # Extract and Rename
-        os.system(f"tar -zxf tasks_compressed/{task_name} && mv tasks/{old_task_name} tasks/{task_name}")
+    if not (os.path.exists(f"tasks/{task_name}")):
+        if not (os.path.exists(f"tasks_compressed/{task_name}")):
+            print(f"Downloading: {task_name}")
+            os.system(f"wget {task_url} -O tasks_compressed/{task_name}")
+        else:
+            print(f"Skipping download of {task_name} - already downloaded")
+
+        print(f"Extracting and renaming: {task_name}")
+        if not old_task_name.startswith("hear-2021"):
+            old_task_name = "tasks/" + old_task_name
+        os.system(
+            f"tar -zxf tasks_compressed/{task_name} && mv {old_task_name} tasks/{task_name}"
+        )
     else:
-        print(f"Skipping download of {task_name} - already downloaded")
+        print(f"{task_name} already prepared")
 
 
 # In[ ]:
-
-
-
-
